requirements.txt
langchain
langchain-community
langsmith
chromadb
tiktoken
ollama
pypdf
streamlit
python-dotenv
watchdog
